Make RMCore receive Node availability requests in parallel to others
Nodes call 'RMCore#setNodeAvailable' periodically. This method returns a
boolean. When a call to 'setNodeAvailable' is handled, a connection
issue which prevents the result to be returned may happen. In such a
situation, it takes several seconds before reaching the protocol
timeout. During this period, no other RMCore methods can be handled.

The purpose of this patch is to handle 'setNodeAvailable' requests on
dedicated threads. This implies to mark the method as Immediate Service
and to ensure that thread-safety is still guaranteed.

It was also noticed that availability is reported per worker on Nodes.
Once done, a check is also done per worker to check if the worker is
still known by the Resource Manager. The patch gathers all method calls
to a single call made to the Resource Manager. In summary, if there were
10 local workers on a Node, a single method call is performed now
instead of 20 before.

This patch resolves ow2-proactive/scheduling#2801.