Prevent killing Nodes JVM when they are detected as down by the RM
Usually, the DefaultInfrastructureManager is used by Nodes that are
deployed with the Agents or manually with the proactive-node script.

A remote down node, attached to a Node Source with
a DefaultInfrastructureManager, must not be removed and killed
immediately since a Node can reconnect by itself. However, a node
removed by a user must be removed and killed immediately.

The solution is far from perfect since the behavior differs from an
InfrastructureManager to another. Unfortunately, it seems to be the best
that can be done without redesigning a lot of parts.

This patch fixes ow2-proactive/scheduling#2798.