Having zero nodes is not an error
Let's assume that a Scheduler and a single remote Node are up and
linked. If a user removes a Node from the RM Web portal, this Node
becomes unknown at the RM level. However, the Node is not immediately
stopped on the Node host. The Node will ask the RM if its URL is still
known, if not, then RMNodeStarter implementation enters the condition
that checks the number of nodes. At this stage, the node count will be
0 but it is not an illegal state.