Forward Node state restoration in NodeSource
The `NodeSource` implementation keeps a view of the ProActive
Programming Nodes it manages. Nodes that are alive and those that are
down are kept in separated data-structures: `nodes` and `downNodes`.

Each NodeSource is responsible to ping the Nodes it knows. Only Nodes
that are alive (i.e. not down) are pinged. I think this is done this
way to prevent a lot of potential timeouts on the RM side (availability
for RM Nodes known as _Down_ in the RM is reported by Nodes themselves).

In `RMCore`, another view is managed. This one keeps track of RMNodes.
A ProActive Programming Node can host zero, one or more RMNodes. At this
level, RMNodes are referenced in 2 different data-structures: one that
lists free nodes and another one that includes all nodes.

When a Node is detected as _Down_, both views are updated. The one in
`NodeSource` but also the one in `RMCore`. However, when an RMNode is
reported as _Free_ (after being detected as _Down_), the `RMCore` view
is updated but not the `NodeSource` view. As a consequence, wrong
Scheduling decisions are made, which slow down Tasks execution.
Furthermore, since a `NodeSource` pings Nodes that are available only,
if the Node experiences a new failure, it will never be detected as
_Down_ again.

This patch fixes ow2-proactive/scheduling#2804.