fix https://github.com/ow2-proactive/studio/issues/277 Studio not usable when authenticated with credential file
Even if this issue occurs on the studio portal, it's root cause is related to the fact that it's not possible to know the user name associated with a rest session generated from a credential file login.

The solution was to allow the retrieval of the authenticated user name from the scheduler and resource manager interfaces, and attach this user name to the rest sessionid.