Handle script variable bindings in ProbabilisticSelectionManager
Selections scripts don't work properly when they use variable bindings such as PA_TASK_REPLICATION

This is because the ProbabilisticSelectionManager caches the result of a script based on the script textual content only, not taking into account variables used inside the script.

This change updates the content of the script with actual variables values before calculating the script DIGEST.

This way, if two tasks use a script with a content such as:

selected = variables.get("PA_TASK_REPLICATION") == 0

When the value of the PA_TASK_REPLICATION differs between the tasks, the digest will be different.

(cherry picked from commit a2800b6)