fix for #2927 RSA decryption randomly fails because of padding issues
Issue was not due to cryptography but to the fact that all replicated tasks shared the same executable container. As replicated task are actually an internal task deeply cloned several times, the test which checks that the container is null always return false, meaning that all replicated tasks share the same container.