Lazy fetch mode for scripts defined by urls  (#3037)
* Lazy fetch mode for scripts defined by urls

When a script is defined by an URL, prior mechanism was to fetch the script immediately at Script instance creation. This change allows to configure the script instance for early or late binding. Late binding means that the url will be resolved at execution time.

Inside the scheduler all scripts will be in lazy mode, except for selection scripts. In case of selection scripts, lazy mode cannot be used as the scheduler server needs the script content during the scheduling loop, prior to the script execution on the nodes.

Also this change adds the possibility to define a script by url and language (when for example, it is not possible to infer the script language from the url value). A new version of the job schema is defined.

 - Refactored HttpClientBuilder > CommonHttpClientBuilder to allow its integration in common-api project.
 - Created CommonHttpResourceDownloader, which makes use of the above class to download http/https resources.
 - Modified FileUtils to extract filename and extension from an url
 - Added a fetchScript method to Script class, this method either return an already defined inline script or try to
 download the defined url, without modifying the internal script object structure.
 - Changed the schemas, xml parser, xml serializer to integrate the <file url="" language="" /> syntax
 - Added in the ScriptData and SelectionScriptData the script url definition
 - When script authorization is enabled, use the fetchScript method to get the script content before authorizing the script.

* Implement reviews