Use the recoveredState handler instead of the frontendState
The frontendState.getState method throws an NPE when accessed from the SchedulerFrontend class. The problem happens during the permission check which is weird because everything is already initialized. If I initialize the monitor runners at a latter stage, I also get the same error which points me to a permission problem. Using the recoveredState fixes the problem, the same object is used by the frontendState (minus the permission check)