Synchronize tark started notification with task execution submitted t… (#3074)
Synchronize tark started notification with task execution submitted to TaskLauncher

The notification and persistence of the task started status was done in parallel with the submission of the task to the TaskLauncher. This small gap led to inconsistencies, especially in the case of a crash of the scheduler within this gap. This commit introduces a small synchronization to be sure to notify the task started status once the doTask request has been received by the TaskLauncher.
- the existing TaskRecoveryWhenNodesAreReservedInBatchTest from commit 018b485c014c48cf8e4edda66931452d10c9aeaa exhibits this particular case.