Fix JobServerLogTest (#3158)
Before it was only job removal but it was bug prone (for pendingJob only) because the following sequence of the events could happen:

SchedulerMethodImpl main loop retrieves info about this forever pending job we call removeJob SchedulerMethodImpl main loop asks for RM::getRMNodes. Third action would print "[SelectionManager.doSelectNodes] "scheduler"..." to the task log. Which would actually re-created removed log file that causes an error in the last line of this method. That is why we kill the job and wait 5s to make sure that iteration of SchedulerMethodImpl main loop is finished, and finally, we remove job with its log files.

As well there is bug fix for RMInitialState