Protect the lambda against null
The lambda expr filling the temporary job was working over a parameter that could have the null value, as a result the entrySet() call would have trown an exception, caught by the putAll() method. My fix is checking the replacementVariables parameter for null, if that's the case, we simply bypass the filling.