Make task pagination O(1) complex (#3262)
When user clicks on the job in scheduling-portal, portal requests server to provide "a page" of 20 tasks of this job. However, on the server side, whole job is copied, and from this job, 20 tasks are copied. So finally portal receives its 20 tasks to show. Because server copies WHOLE job (with all its tasks), complexity of this operation is linear to number of tasks in the job.

This PR will make complexity constant by copying only requested tasks.