Allow timeout on Authentication.isActivated
The previous implementation used a ProActive synchronous call (primitive boolean return value) which can block indefinitely. This could cause random freeze of a windows agent trying to contact the RM (the precise reason why the method never return is still uncertain).

This modification allow to not wait more than the expected timeout, allowing retries on windows agents.

An immediate service call is also added to try to avoid the root cause of the issue.