LoadPackage : create new revision when upgrading the scheduler
When an existing installation of the scheduler is upgraded to a new version, the current behavior is that all catalog objects present already before the upgrade are not modified.

This is a problem as the new version may bring new improvements of the catalog workflows, patches, etc.

This commit changes the behavior and creates a new revision of the existing catalog object.

The administrator can then decide using the catalog portal if he wants to keep the update or rollback if necessary.