Stop the scheduler only in case of jython deadlocks
It has been observed recently that shutdown of the scheduler server occurs in case of server lag, not due to python deadlocks. The goal of this change is to trigger a shutdown only when the cause is due to jython and not something else.

(cherry picked from commit 1ba5c072897de5696ab8be777b72d0da70db9bf8)