Remove and comment recursive algorithm in InternalJob
Replaced by a simple iterative search

After analysis of the code, it tries to find an InternalTask based on its name, but to do so, it does a complex  potentially exponential recursive algorithm. This code hanged a task termination on a workflow containing a loop which had around 1400 tasks.

For now I comment the previous code to keep its track.