Disable scheduling loop node partitioning
Even though in general usage the current implementation works, it became very inefficient when there are nodes protected by tokens. In that case, if many tasks (10K more) needs to be scheduled, the scheduling loop is querying the rm several thousand times for a small number of nodes.

This change disable the node partitioning. When there are 10K tasks to be scheduled and all are "compatible", the scheduling loop send a single request to the RM, which returns in a single call the available nodes.

(cherry picked from commit e12fd333590f63830b0844d44ff7330c81a75178)