Add JPCanCreateAssociationPermission
This permission is aimed to be checked by the job-planner service (and portal) to ensure that the user has the right to create an association

Accordingly, added this permission to all users that have access to the job-planner portal.
Can be disabled for read-only users (e.g. guests).

(cherry picked from commit 8d2d3711d706bbbdd8b41e65b6d52ec3c1cd3397)