Fix updateStartAt
When updating the startAt value of a pending job, the cache used by ExtendedSchedulerPolicy need to be updated.

Accordingly, the StartAtUpdater must be able to retrieve the policy instance and ask it to update the startAt value of a given job or task.

(cherry picked from commit 5f58dbfa667752769d52efa22bb2ae5f41e7bd98)