Store script urls in database by reference
When this option is enabled, script urls which contains a catalog url will be stored in the db by using $PA_CATALOG_REST_URL/path and this expression will be resolved to a local url when the job is loaded from the db.

To use this option, it is mandatory to set the catalog url manually in config/scheduler/settings.ini.

This is to prevent issues in HA active/passive of two servers using the same database

(cherry picked from commit ef1204fdaaf054a6e4e906f358e5b009c0da9558)